Instructions
1) Program is in Java
2) Need to have IntelliJ installed
3) Amazon Corretto JDK 11 installed
4) Open the three terminals. Navigate to --> ups_FinalProject --> out --> production
5) Run the command --> java rpsServer
6) Run two copies in separate terminals --> java rpsClient 

As listed above there are two things you must have installed in order to run this project. But in order for the project to run correctly you must first get the rpsServer running first and then run two copies of rpsClient via the terminal. The reason for this is because the project will not in IntelliJ by just building and running it. Due to needing some command line arguments that will be assigned to the IP address and the port number variables. So make sure that you run this in the terminals or manually open the files and hard code those values. 

[Note: When playing the game you must enter the first clients choice first before entering the second clients choice. So the project will crash I was not able to get the server to manage this client. But as long as your send the first connected clients choice first the game will run smoothly.]